requirements.txt
aniso8601==4.1.0
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.1
Click==7.0
cryptography==2.5
Flask==1.0.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.9.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==3.0.0rc4
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.19
PyJWT==1.7.1
pytz==2018.9
redis==3.2.0
six==1.12.0
SQLAlchemy==1.2.18
Werkzeug==0.14.1
